requirements.txt
flask==2.3.3
requests==2.31.0
beautifulsoup4==4.12.2
selenium==4.15.0
plotly==5.17.0
dash==2.14.2
dash-bootstrap-components==1.5.0
dash-cytoscape==0.3.0
pandas==2.1.3
networkx==3.2.1
phonenumbers==8.13.25
email-validator==2.1.0
python-whois==0.8.0
shodan==1.30.1
tweepy==4.14.0
praw==7.7.1
linkedin-api==2.0.0
googlesearch-python==1.2.3
aiohttp==3.9.1
asyncio==3.4.3
fake-useragent==1.4.0
lxml==4.9.3
python-dotenv==1.0.0
jinja2==3.1.2
werkzeug==2.3.7
# Added for real data integrations
dnspython==2.6.1
python-whois==0.8.0
socialscan==1.4.2
requests==2.31.0
